refactor(api):rename model file name as the model name
