+++ api/tests/unit/domain/usecases/create-organization_test.js
@@ -1,6 +1,6 @@
 import { expect, sinon, catchErr } from '../../../test-helper.js';
 import { createOrganization } from '../../../../lib/domain/usecases/create-organization.js';
-import { OrganizationForAdmin } from '../../../../lib/domain/models/organizations-administration/Organization.js';
+import { OrganizationForAdmin } from '../../../../lib/domain/models/organizations-administration/OrganizationForAdmin.js';
 import { EntityValidationError } from '../../../../lib/domain/errors.js';
 
 describe('Unit | UseCase | create-organization', function () {
