+++ api/tests/unit/domain/usecases/organizations-administration/update-organization-information_test.js
@@ -2,7 +2,7 @@ import { expect, sinon, catchErr, domainBuilder } from '../../../../test-helper.
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { Tag } from '../../../../../lib/domain/models/Tag.js';
 import { OrganizationTag } from '../../../../../lib/domain/models/OrganizationTag.js';
-import { OrganizationForAdmin } from '../../../../../lib/domain/models/organizations-administration/Organization.js';
+import { OrganizationForAdmin } from '../../../../../lib/domain/models/organizations-administration/OrganizationForAdmin.js';
 import * as OidcIdentityProviders from '../../../../../lib/domain/constants/oidc-identity-providers.js';
 import { usecases } from '../../../../../lib/domain/usecases/index.js';
 const { updateOrganizationInformation } = usecases;
